Navigation bar

+++ src/mml/Model/Movie.java
@@ -2,6 +2,7 @@ package mml.Model;
 
 import javax.swing.*;
 import java.awt.*;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -35,6 +36,8 @@ public class Movie {
     private String Production;
     private String Website;
 
+    transient private ImageIcon posterIcon;
+
     public String getTitle(){
         return Title;
     }
@@ -45,12 +48,6 @@ public class Movie {
         return Year;
     }
 
-    /*public List<String> getActors(){
-        if (Actors.equals("N/A")){
-            return null;
-        }
-        return Arrays.asList(Actors.split("\\s*,\\s*"));
-    }*/
     public String getActors() { return Actors; };
 
     public List<String> getDirector(){
@@ -86,37 +83,48 @@ public class Movie {
         return Ratings;
     }
 
-    public String getPoster() { return Poster; }
+    public ImageIcon getPoster() {
+        return posterIcon;
+    }
 
     public String getReleased() { return Released; }
 
     public String getLanguage() { return Language; }
 
-    public ImageIcon createPoster(String posterURL){
-        ImageIcon posterIcon;
-        try {
-            String destinationFile = "src/poster.jpg";
-            URL url = new URL(posterURL);
-            InputStream is = url.openStream();
-            OutputStream os = new FileOutputStream(destinationFile);
+    public void createPoster(){
+        File posterFile = new File("src/Images/Posters/" + Title);
+        if (posterFile.exists()){
+            posterIcon = new ImageIcon(new ImageIcon(posterFile.getPath()).getImage()
+                    .getScaledInstance(300, 445, Image.SCALE_DEFAULT));
+        }
+        else {
+            try {
+                String destinationFile = "src/Images/Posters/" + Title;
+                URL url = new URL(Poster);
+                InputStream is = url.openStream();
+                OutputStream os = new FileOutputStream(destinationFile);
 
-            byte[] b = new byte[2048];
-            int length;
+                byte[] b = new byte[2048];
+                int length;
 
-            while ((length = is.read(b)) != -1) {
-                os.write(b, 0, length);
+                while ((length = is.read(b)) != -1) {
+                    os.write(b, 0, length);
+                }
+
+                is.close();
+                os.close();
+
+                posterIcon = new ImageIcon(new ImageIcon(destinationFile).getImage()
+                        .getScaledInstance(300, 445, Image.SCALE_DEFAULT));
             }
-
-            is.close();
-            os.close();
-
-            posterIcon = new ImageIcon(new ImageIcon("src/poster.jpg").getImage()
-                    .getScaledInstance(300, 445, Image.SCALE_DEFAULT));
+            catch (Exception e){
+                posterIcon = new ImageIcon(new ImageIcon("src/Images/defaultPoster.jpg").getImage()
+                        .getScaledInstance(300, 445, Image.SCALE_DEFAULT));
+            }
         }
-        catch (Exception e){
-            posterIcon = new ImageIcon(new ImageIcon("src/defaultPoster.jpg").getImage()
-                    .getScaledInstance(300, 445, Image.SCALE_DEFAULT));
-        }
-        return posterIcon;
+    }
+
+    public String getImdbRating(){
+        return imdbRating;
     }
 }
